pycodestyle

+++ 0x02-python_async_comprehension/1-async_comprehension.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python3
 
 """async comprenhension"""
-
-import random
+from asyncio import sleep
+from random import uniform
 from typing import List
 async_generator = __import__('0-async_generator').async_generator
 
 
-async def async_comprehension() -> List[int]:
+async def async_comprehension() -> List[float]:
     """
     coroutine will collect 10 random ints over async_generator
     """
